requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
openai==1.3.0
pydantic==2.5.0
python-dotenv==1.0.0
moviepy==1.0.3
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
aiofiles==23.2.1
websockets==12.0
